tidy up Fd code - move functionality to retrieve xdp statistics there and remove ability to clone and copy to ensure it doesn't outlive the socket
